+++ restler-core/src/main/java/net/researchgate/restdsl/metrics/NoOpMetricSink.java
@@ -1,0 +1,10 @@
+package net.researchgate.restdsl.metrics;
+
+public class NoOpMetricSink implements MetricSink {
+    public static final NoOpMetricSink INSTANCE = new NoOpMetricSink();
+
+    @Override
+    public void timing(MetricName name, long durationMs) {
+
+    }
+}