+++ restler-core/src/main/java/net/researchgate/restdsl/metrics/NoOpStatsReporter.java
@@ -2,7 +2,11 @@ package net.researchgate.restdsl.metrics;
 
 /**
  * No op
+ *
+ * @deprecated Use {@link NoOpMetricSink}
  */
+@Deprecated(since = "6.1.0", forRemoval = true)
+@SuppressWarnings("removal")
 public class NoOpStatsReporter implements StatsReporter {
     public static final NoOpStatsReporter INSTANCE = new NoOpStatsReporter();